feat: 카테고리 대시보드 전체 조회 비즈니스 로직 구현

+++ src/main/java/com/pinback/pinback_server/domain/category/presentation/dto/response/CategoryDashboardResponse.java
@@ -3,6 +3,6 @@ package com.pinback.pinback_server.domain.category.presentation.dto.response;
 public record CategoryDashboardResponse(
 	long categoryId,
 	String categoryName,
-	int unreadCount
+	long unreadCount
 ) {
 }
